+++ lib/brcobranca/boleto/base.rb
@@ -91,6 +91,8 @@ module Brcobranca
       attr_accessor :cedente_endereco
       # <b>OPCIONAL</b>: EMV para gerar QRCode para pagamento via PIX
       attr_accessor :emv
+      # <b>OPCIONAL</b>: Label exibido ao lado do QR Code PIX (default: "Pague com PIX")
+      attr_accessor :pix_label
       # <b>OPCIONAL</b>: Descontos e abatimentos
       attr_accessor :descontos_e_abatimentos